requirements.txt
torch
numpy
ipywidgets
opencv-python
matplotlib
scikit-learn
ipykernel
scikit-image
pandas
pyyaml